requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
streamlit==1.32.0
langchain==0.3.4
langchain-core==0.3.12
langchain-community==0.3.3
langchain-ollama==0.2.0
pydantic==2.9.2
requests==2.32.3
PyMuPDF==1.24.11
gTTS==2.5.3
speechrecognition==3.10.4
streamlit-mermaid==0.2.0
tenacity==8.2.3
bcrypt==4.2.0
python-dotenv==1.0.1
aiofiles==24.1.0
sentence-transformers==3.0.1
torch==2.3.1 
transformers==4.43.3
scipy==1.14.0
pandas==2.2.2
google-search-results==2.4.2
streamlit-webrtc==0.47.0
pyaudio==0.2.14
pymilvus==2.4.9